refactor(run): scope walk moves to its own module — loc-limits holds
mod.rs crossed the 1500-LOC law by seven lines of documentation; the
--task ancestor-cone walk (pure graph · no lane logic) was the natural
tenant to rehome. Zero behavior change · 196 lib tests green.
