+++ crates/nika-cli/src/verbs/run/scope.rs
@@ -1,0 +1,57 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+// Copyright (C) 2024-2026 SuperNovae Studio <contact@supernovae.studio>
+
+//! `--task` scoping — the ancestor-cone cut behind the regenerate-one-
+//! block move (its gate + re-check live in the parent module; this file
+//! is the pure graph walk).
+
+use nika_schema::raw::RawWorkflow;
+
+/// Scope a workflow to ONE task + its transitive upstream (`--task`).
+///
+/// Ancestors must run — their outputs feed the target's bindings; nothing
+/// downstream or sibling executes. Document order is preserved (stable
+/// waves) and workflow `outputs:` drop (they may reference tasks outside
+/// the scope — the target's own output IS the point of the run). Unknown
+/// ids fail with the available set (environment class · exit 3 · before
+/// any effect — the same lane as an unknown `--var` key).
+pub(super) fn scope_to_task(mut wf: RawWorkflow, target: &str) -> Result<RawWorkflow, String> {
+    use std::collections::{BTreeSet, VecDeque};
+
+    let mut deps_of: std::collections::BTreeMap<String, Vec<String>> =
+        std::collections::BTreeMap::new();
+    for t in &wf.tasks {
+        deps_of.insert(
+            t.value.id.value.as_str().to_owned(),
+            t.value
+                .depends_on
+                .iter()
+                .map(|d| d.value.as_str().to_owned())
+                .collect(),
+        );
+    }
+    if !deps_of.contains_key(target) {
+        let known = deps_of.keys().cloned().collect::<Vec<_>>().join(" · ");
+        return Err(format!(
+            "--task `{target}` names no task in this workflow — tasks: {known}"
+        ));
+    }
+
+    let mut keep: BTreeSet<String> = BTreeSet::new();
+    let mut queue: VecDeque<String> = VecDeque::from([target.to_owned()]);
+    while let Some(id) = queue.pop_front() {
+        if !keep.insert(id.clone()) {
+            continue;
+        }
+        if let Some(deps) = deps_of.get(&id) {
+            for d in deps {
+                queue.push_back(d.clone());
+            }
+        }
+    }
+
+    wf.tasks
+        .retain(|t| keep.contains(t.value.id.value.as_str()));
+    wf.outputs.clear();
+    Ok(wf)
+}
